requirements-dev.txt
aws-cdk-lib
constructs
pytest==6.2.5
boto3
fastapi
gradio
pandas
tabulate
SpeechRecognition